+++ R/comparisons.R
@@ -245,7 +245,7 @@ comparisons <- function(model,
         # if `newdata` is a call to `datagrid`, `typical`, or `counterfactual`,
         # insert `model` should probably not be nested too deeply in the call
         # stack since we eval.parent() (not sure about this)
-        scall <- substitute(newdata)
+        scall <- rlang::enquo(newdata)
         newdata <- sanitize_newdata_call(scall, newdata, model)
 
         model <- sanitize_model(
@@ -635,7 +635,7 @@ avg_comparisons <- function(model,
 
     # order of the first few paragraphs is important
     # if `newdata` is a call to `typical` or `counterfactual`, insert `model`
-    scall <- substitute(newdata)
+    scall <- rlang::enquo(newdata)
     newdata <- sanitize_newdata_call(scall, newdata, model)
 
     # Bootstrap