+++ R/plot_slopes.R
@@ -72,7 +72,7 @@ plot_slopes <- function(model,
     # order of the first few paragraphs is important
     # if `newdata` is a call to `typical` or `counterfactual`, insert `model`
     # should probably not be nested too deeply in the call stack since we eval.parent() (not sure about this)
-    scall <- substitute(newdata)
+    scall <- rlang::enquo(newdata)
     newdata <- sanitize_newdata_call(scall, newdata, model)
 
     valid <- c("dydx", "eyex", "eydx", "dyex")