+++ inst/tinytest/test-newdata.R
@@ -77,6 +77,23 @@ expect_inherits(slopes(mod, newdata = dat, by = "cyl"), "slopes")
 # expect_equivalent(sort(cmp$std.error), sort(emm$SE))
 
 
+# Issue #814
+data(lalonde, package = "MatchIt")
+if(exists("mdata")) rm(mdata)
+test <- function() {
+    mdata <- lalonde
+    m0 <- lm(re78 ~ nodegree, data = mdata)
+    comparisons(m0, variables = "nodegree", newdata = subset(mdata, nodegree == 1))
+}
+cmp1 <- test()
+mdata <- subset(lalonde, married == 1)
+m0 <- lm(re78 ~ nodegree, data = mdata)
+cmp2 <- comparisons(m0, variables = "nodegree", newdata = subset(mdata, nodegree == 1))
+cmp3 <- test()
+expect_equal(nrow(cmp1), nrow(subset(lalonde, nodegree == 1)))
+expect_equal(nrow(cmp2), nrow(subset(lalonde, nodegree == 1 & married == 1)))
+expect_equal(nrow(cmp3), nrow(subset(lalonde, nodegree == 1)))
+
 
 
 rm(list = ls())